transfer code from rewrite into other files
