+++ Project.toml
@@ -12,7 +12,6 @@ HCIToolbox = "b6cd55e5-4d02-4e12-b82c-005f67e784bf"
 ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 NMF = "6ef6ca0d-6ad7-5ff6-b225-e928bfa0a386"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Photometry = "af68cb61-81ac-52ed-8703-edc140936be4"
 ProgressLogging = "33c8b6b6-d38a-422a-b730-caa89a2f386c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -24,7 +23,6 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 TSVD = "9449cd9e-2762-5aa3-a617-5413e99d722e"
-UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
 CoordinateTransformations = "0.5, 0.6"
@@ -34,7 +32,6 @@ FillArrays = "0.8, 0.9, 0.10"
 HCIToolbox = "0.5.1"
 ImageTransformations = "0.8"
 NMF = "0.4.1, 0.5"
-Parameters = "0.12"
 Photometry = "0.7"
 ProgressLogging = "0.1"
 Reexport = "0.2, 1"
@@ -45,5 +42,4 @@ StaticKernels = "0.5, 0.6"
 StatsBase = "0.33"
 StructArrays = "0.4"
 TSVD = "0.4"
-UnPack = "1"
 julia = "1.4"
