+++ src/greeds.jl
@@ -2,52 +2,87 @@ using ProgressLogging
 using Setfield
 
 """
-    GreeDS(alg=PCA(); threshold=0.0, progress=true)
-    GreeDS(ncomps; threshold=0.0, progress=true)
+    GreeDS(alg=PCA(); threshold=0.0)
+    GreeDS(ncomps; threshold=0.0, options...)
 
 Performs the greedy disk subtraction (GreeDS) algorithm.
 
-This method is an iterative approach to standard ADI reduction which seeks to minimize over-subtraction by constraining the low-rank matrix approximation from `alg`.
+This method is an iterative approach to standard ADI reduction which seeks to minimize over-subtraction by constraining the low-rank matrix approximation from `alg`. By default, uses [`PCA`](@ref). If `ncomps` or other PCA options are provided, they will be passed to the constructor.
 
 For large data cubes the iteration can cause slowdowns, so a progress bar is provided using the [`ProgressLogging`](https://github.com/JunoLab/ProgressLogging.jl) API along with the `progress` keyword. It won't appear without a logging backend, such as [`TerminalLoggers`](https://github.com/c42f/TerminalLoggers.jl).
 
 # Algorithms
-Although the original paper explicitly uses PCA, we allow use of any linear ADI algorithm that is characterized by `ncomps`. By default, uses [`PCA`](@ref). If `ncomps` are provided, will default to [`PCA`](@ref) as well.
+Originally multiple algorithms were supported, but currently only [`PCA`](@ref) works properly with the GreeDS algorithm. 
 
 # References
 1. [Pairet et al. 2018](https://ui.adsabs.harvard.edu/abs/2018arXiv181201333P) "Reference-less algorithm for circumstellar disks imaging"
 2. [Pairet et al. 2020](https://ui.adsabs.harvard.edu/abs/2020arXiv200805170P) "MAYONNAISE: a morphological components analysis pipeline for circumstellar disks and exoplanets imaging in the near infrared"
 """
-struct GreeDS{ALG<:LinearAlgorithm} <: LinearAlgorithm
-    alg::ALG
-    threshold::Float64
-    progress::Bool
+struct GreeDS{AT} <: ADIAlgorithm
+    kernel::AT
+    threshold
+end
+GreeDS(alg=PCA(); threshold=0.0) = GreeDS(alg, threshold)
+GreeDS(ncomps::Int; threshold=0.0, kwargs...) = GreeDS(PCA(ncomps; kwargs...), threshold=threshold)
+
+function (alg::GreeDS)(cube, angles; method=:deweight, kwargs...)
+    return collapse!(subtract(alg, cube; angles=angles, kwargs...), angles, method=method)
 end
 
-GreeDS(alg=PCA(); threshold=0.0, progress=true) = GreeDS(alg, threshold, progress)
-GreeDS(ncomps::Int; kwargs...) = GreeDS(PCA(ncomps); kwargs...)
-
-function decompose(alg::GreeDS, cube, angles, cube_ref=cube; kwargs...)
-    target = expand(cube_ref)
+function fit(alg::GreeDS{<:Union{PCA,TPCA}}, data::AbstractArray{T,3}; angles, ref=data) where T
+    target = flatten(ref)
     # get the number of components as a range from the underlying alg
-    max_ncomps = isnothing(alg.alg.ncomps) ? size(target, 1) : alg.alg.ncomps
+    max_ncomps = get_ncomps(alg.kernel.ncomps, target)
     # use the underlyhing algorithm with a lens for further processing
-    _alg = alg.alg
-    _alg = @set _alg.ncomps = 1
-    reduced = _alg(target, angles)
-    local A
+    f = alg.kernel
+    f = @set f.ncomps = 1
+    design = fit(f, target)
+    R = expand(target .- reconstruct(design))
+    reduced = collapse!(R, angles)
     @progress "GreeDS" for n in 1:max_ncomps
-        resid = target .- expand_rotate(reduced, angles, alg.threshold; kwargs...)
+        resid = data .- expand_rotate(reduced, angles, alg.threshold)
         # use lens to update number of components
-        _alg = @set _alg.ncomps = n
+        f = @set f.ncomps = n
         # use the `resid` cube as the reference frames for the next reduction
-        A, w = decompose(_alg, target, angles, resid; kwargs...)
-        reduced = collapse!(target .- expand(w * A), angles; kwargs...)
+        design = fit(f, target; ref=flatten(resid))
+        R = expand(target .- reconstruct(design))
+        reduced = collapse!(R, angles)
     end
-    # project target cube onto reference basis
-    w = flatten(cube) * A'
-    return A, w
+    if ref !== data
+        A = design.components
+        design.weights .= flatten(data) * A'A
+    end
+    return design
 end
+
+function fit(alg::GreeDS{<:Union{PCA,TPCA}}, data::AnnulusView; angles, ref=data)
+    target = data()
+    tmpAnn = copy(data)
+    # get the number of components as a range from the underlying alg
+    max_ncomps = get_ncomps(alg.kernel.ncomps, target)
+    # use the underlyhing algorithm with a lens for further processing
+    f = alg.kernel
+    f = @set f.ncomps = 1
+    design = fit(f, target)
+    R = inverse(data, target .- reconstruct(design))
+    reduced = collapse!(R, angles)
+    @progress "GreeDS" for n in 1:max_ncomps
+        tmpAnn.parent .= data .- expand_rotate(reduced, angles, alg.threshold)
+        resid = tmpAnn()
+        # use lens to update number of components
+        f = @set f.ncomps = n
+        # use the `resid` cube as the reference frames for the next reduction
+        design = fit(f, target; ref=resid)
+        R = inverse(data, target .- reconstruct(design))
+        reduced = collapse!(R, angles)
+    end
+    if ref !== data
+        A = design.components
+        design.weights .= flatten(data) * A'A
+    end
+    return design
+end
+
 
 """
     expand_rotate(frame, angles, threshold; kwargs...)
