+++ src/nmf.jl
@@ -1,4 +1,5 @@
-using NMF: nnmf, solve!, CoordinateDescent, nndsvd
+
+using NMF: nndsvd, solve!, CoordinateDescent
 
 """
     NMF(;ncomps=nothing) <: LinearAlgorithm
@@ -16,67 +17,78 @@ If `ncomps` is `nothing`, it will be set to the number of frames in the referenc
     ```
     When doing full-frame reduction (e.g. `NMF()(cube, angles)`) *this is handled automatically*, so this constraint only applies to the lower-level API and methods which rely on those, such as [`GreeDS`](@ref). **In general, if you see warnings, heed them.**
 
-# Implements
-* [`decompose`](@ref)
+!!! danger "RDI"
+    RDI is not currently supported for NMF due to upstream limitations in NMF.jl
 
 # References
 1. [Ren et al. 2018](http://adsabs.harvard.edu/abs/2018ApJ...852..104R) Non-negative Matrix Factorization: Robust Extraction of Extended Structures
 """
-@with_kw struct NMF <: LinearAlgorithm
-    ncomps::Union{Int,Nothing} = nothing
+struct NMF <: ADIAlgorithm
+    ncomps
 end
+NMF(; ncomps=nothing) = NMF(ncomps)
 
-# manually check for negatives and rescale for full-frame processing
-function (alg::NMF)(cube::AbstractArray{T,3}, angles; kwargs...) where T
-    target = any(<(0), cube) ? cube .- minimum(cube) : cube 
-    S = reconstruct(alg, target, angles)
-    residual_cube = target .- S
-    return collapse!(residual_cube, angles; kwargs...)
-end
-
-# manually check for negatives and rescale for full-frame processing
-function (alg::NMF)(cube::AbstractArray{T,3}, angles, cube_ref; kwargs...) where T
+function process(alg::NMF, cube; kwargs...)
     if any(<(0), cube)
-        minval = min(minimum(cube), minimum(cube_ref))
-        target = cube .- minval
-        ref = cube_ref .- minval
+        target = copy(cube)
+        target .-= minimum(target)
     else
         target = cube
-        ref = cube_ref
     end
-    S = reconstruct(alg, target, angles, ref)
-    residual_cube = target .- S
-    return collapse!(residual_cube, angles; kwargs...)
+    S = reconstruct(alg, target; kwargs...)
+    return target .- S
 end
 
-# TODO this doesn't quite match what sklearn does
-function ADI.decompose(alg::NMF, cube, angles, cube_ref; kwargs...)
-    if any(<(0), cube) || any(<(0), cube_ref)
-        @warn "Negative values encountered in `cube` or `cube_ref`. Make sure to rescale your inputs"
-    end
-    X = flatten(cube)
-    X_ref = flatten(cube_ref)
-
-    k = isnothing(alg.ncomps) ? size(cube, 1) : alg.ncomps
-    k > size(cube, 1) && error("ncomps ($k) cannot be greater than the number of frames ($(size(cube, 1)))")
-    # fit H using reference
-    _, H = nndsvd(X_ref, k)
-    W = X * H'
-    solve!(CoordinateDescent{eltype(X)}(), X, W, H)
-    return H, W
+struct NMFDesign{AT,WT} <: LinearDesign
+    components::AT
+    weights::WT
 end
 
-function ADI.decompose(alg::NMF, cube, angles; kwargs...)
-    # X = clamp.(flatten(cube), 0, Inf) # clamp non-negative values
-    if any(<(0), cube)
-        @warn "Negative values encountered in `cube`. Make sure to rescale your inputs"
+design(des::NMFDesign) = (des.components, des.weights)
+
+function fit(alg::NMF, data::AbstractMatrix{T}) where T
+    if any(<(0), data)
+        @warn "Negative values encountered in input. Make sure to rescale your data"
     end
-    X = flatten(cube)
+    if alg.ncomps === nothing
+        k = size(data, 1)
+    else
+        k = min(alg.ncomps, size(data, 1))
+    end
 
-    k = isnothing(alg.ncomps) ? size(cube, 1) : alg.ncomps
-    k > size(cube, 1) && error("ncomps ($k) cannot be greater than the number of frames ($(size(cube, 1)))")
-
-    W, H = nndsvd(X, k)
-    solve!(CoordinateDescent{eltype(X)}(), X, W, H)
-    return H, W
+    W, H = nndsvd(data, k)
+    solve!(CoordinateDescent{T}(), data, W, H)
+    return NMFDesign(H, W)
 end
+
+# # TODO this doesn't quite match what sklearn does
+# function ADI.decompose(alg::NMF, cube, angles, cube_ref; kwargs...)
+#     if any(<(0), cube) || any(<(0), cube_ref)
+#         @warn "Negative values encountered in `cube` or `cube_ref`. Make sure to rescale your inputs"
+#     end
+#     X = flatten(cube)
+#     X_ref = flatten(cube_ref)
+
+#     k = isnothing(alg.ncomps) ? size(cube, 1) : alg.ncomps
+#     k > size(cube, 1) && error("ncomps ($k) cannot be greater than the number of frames ($(size(cube, 1)))")
+#     # fit H using reference
+#     _, H = nndsvd(X_ref, k)
+#     W = X * H'
+#     solve!(CoordinateDescent{eltype(X)}(), X, W, H)
+#     return H, W
+# end
+
+# function ADI.decompose(alg::NMF, cube, angles; kwargs...)
+#     # X = clamp.(flatten(cube), 0, Inf) # clamp non-negative values
+#     if any(<(0), cube)
+#         @warn "Negative values encountered in `cube`. Make sure to rescale your inputs"
+#     end
+#     X = flatten(cube)
+
+#     k = isnothing(alg.ncomps) ? size(cube, 1) : alg.ncomps
+#     k > size(cube, 1) && error("ncomps ($k) cannot be greater than the number of frames ($(size(cube, 1)))")
+
+#     W, H = nndsvd(X, k)
+#     solve!(CoordinateDescent{eltype(X)}(), X, W, H)
+#     return H, W
+# end
